requirements.txt
azure-functions
oci
requests
msrestazure
azure-loganalytics